fix: remove \n before input- and output- prompt text (#6)
Input- and output- prompts are the "In [3]:" and "Out [3]:" bits
next to code cells and code cell outputs.

In general: added a way to control if a newline should be added
after a tag or not. This is a dirty solution, but that's its only
real drawback at the moment.

v0.3.0 will ship with a decent rewrite of the HTML/CSS logic,
so I am going to postpone implementing a better solution until then.